Move depot UI to separate files
