+++ src/window/building_info.c
@@ -33,6 +33,7 @@
 #include "window/advisors.h"
 #include "window/city.h"
 #include "window/message_dialog.h"
+#include "window/building/depot.h"
 #include "window/building/distribution.h"
 #include "window/building/figures.h"
 #include "window/building/government.h"
